requirements.txt
Django==5.1.3
redis==5.2.0
celery==5.4.0
pydantic==2.10.1
gunicorn==23.0.0
django-extensions=3.2.3
django-celery-beat==2.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1